+++ test/with-typescript/index.ts
@@ -2019,10 +2019,6 @@ ThirdPartyPasswordless.consumePasswordlessCode({
         preAPIHook: undefined,
     },
 });
-// @ts-expect-error
-ThirdPartyPasswordless.consumePasswordlessCode(undefined);
-// @ts-expect-error
-ThirdPartyPasswordless.consumePasswordlessCode();
 
 ThirdPartyPasswordless.createPasswordlessCode({
     email: "",